Rezept direkt nach KI-Analyse importieren statt Zwischenschritt anzuzeigen
URL-Import, Foto-OCR-Scan und der Universal-Import (Web-Share-Target)
zeigten nach der KI-Analyse einen reinen Lesevorschau-Screen, bevor der
Nutzer per Klick auf "Übernehmen" ins (ohnehin voll editierbare)
Rezeptformular gelangte. Dieser Zwischenschritt war technisch nicht
notwendig - er diente nur als Bestätigungs-Gate. Die Ergebnisse werden
jetzt direkt nach Abschluss der Analyse importiert, der Nutzer landet
sofort im editierbaren Formular.

Die gemeinsame Recipe-Objekt-Erstellung aus einem KI-Ergebnis wurde nach
buildRecipeFromAiResult() in ocrParser.js ausgelagert, da sie zuvor
dreifach dupliziert war.
